refactor: ログ機能をmain.pyから分離
